requirements.txt
# Core Python dependencies
aiohttp
python-dotenv
gradio
pandas
pathlib
huggingface_hub

# Logging and utilities

jsonschema
typing-extensions

# Development and testing (optional)
# pytest
# mypy

# API and LLM interaction
# OpenRouter API is used, but no specific library needed beyond aiohttp